Remove src/neo/SmartContract/Nns/NnsContract.RegisterCenter.cs
